+++ plugins/speech/scratch.js
@@ -12,4 +12,6 @@ txt = 'seventh millenium. white collar'
 // let doc = nlp(txt).compute(['soundsLike', 'syllables'])
 // console.dir(doc.json()[0], { depth: 5 })
 
-console.log(nlp(txt).syllables())
+let doc = nlp('calgary')
+doc.compute('soundsLike')
+console.log(JSON.stringify(doc.json()[0], null, 2))