+++ plugins/speech/src/api.js
@@ -14,5 +14,20 @@ const api = function (View) {
     })
     return all
   }
+  /** */
+  View.prototype.soundsLike = function () {
+    this.compute('soundsLike')
+    let all = []
+    this.docs.forEach(terms => {
+      let some = []
+      terms.forEach(term => {
+        some = some.concat(term.soundsLike)
+      })
+      if (some.length > 0) {
+        all.push(some)
+      }
+    })
+    return all
+  }
 }
 export default api